# openapi/openapi.yaml
openapi: "3.0.3"
info:
  title: K8sPenTool-ng API
  description: |
    K8sPenTool-ng - Kubernetes Container Penetration Testing Platform

    Integrates CDK (Container Defense Kit) exploits with AI-driven attack chain automation.
    All endpoints use JSON request/response. Real-time events via WebSocket at `/api/v1/ws`.
  version: "2.0.0"
  contact:
    name: K8sPenTool Team
    url: https://github.com/trymonoly/K8sPenTool-ng
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
...
paths:
...
  /access/kubelet/exec:
    post:
      tags: [Access]
      summary: Execute command via Kubelet API
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KubeletExecRequest'
      responses:
        '200':
          description: Command output

  # Access - Etcd
...
  /exec/api-server/exec:
    post:
      tags: [Exec]
      summary: Execute command in pod via APIServer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PodExecRequest'
      responses:
        '200':
          description: Command output
...
  /exec/kubelet/exec:
    post:
      tags: [Exec]
      summary: Execute command via Kubelet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KubeletExecRequest'
      responses:
        '200':
          description: Command output
...
  /kubectl/exec:
    post:
      tags: [Kubectl]
      summary: Execute arbitrary kubectl command
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KubectlCommandRequest'
      responses:
        '200':
          description: Command output
...
components:
  schemas:
...
    KubeletExecRequest:
      type: object
      required: [target_host, pod_name, command]
      properties:
        target_host:
          type: string
        namespace:
          type: string
        pod_name:
          type: string
        container_name:
          type: string
        command:
          type: string
        token:
          type: string
        timeout_sec:
          type: integer
        skip_tls:
          type: boolean
...
    PodExecRequest:
      type: object
      required: [target_host, pod_name, command]
      properties:
        target_host:
          type: string
        namespace:
          type: string
        pod_name:
          type: string
        container_name:
          type: string
        command:
          type: string
        token:
          type: string
        username:
          type: string
        password:
          type: string
        timeout_sec:
          type: integer
        skip_tls:
          type: boolean
...
    KubectlRequest:
      type: object
      required: [target_host]
      properties:
        target_host:
          type: string
        token:
          type: string
        username:
          type: string
        password:
          type: string
        skip_tls:
          type: boolean
        kubeconfig_path:
          type: string
        kubeconfig:
          type: string
        timeout_sec:
          type: integer
    KubectlCommandRequest:
      allOf:
        - $ref: '#/components/schemas/KubectlRequest'
        - type: object
          required: [command]
          properties:
            command:
              type: string
...